+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * Add `WP_STASH_BYPASS` environment variable
 * object-cache.php drop-in now merely requires the actual drop-in from the install folder. This makes it update-safe
 * Add a couple of type assertion for improved IDE support in object-cache.php
+* Use Composite Driver for staggered caching instead of homebrew local memory cache.
+
 #### 3.1.0
 * Implementation of `Debug\ActionLogger`
 * Added new Logger to `Pool`